PR comments - dimensions -> dimensionValues
