+++ server/collections/resources/dimensionValues/us_pa.js
@@ -22,50 +22,39 @@ const {
   DEFAULT_RACE,
 } = require("./shared");
 
-const DEFAULT_CHARGE_CATEGORIES = [
+const DEFAULT_SUPERVISION_LEVELS = [
   "all",
-  "domestic_violence",
-  "general",
-  "serious_mental_illness",
-  "sex_offense",
+  "electronic_monitoring_only",
+  "enhanced",
   "external_unknown",
+  "maximum",
+  "medium",
+  "minimum",
+  "special",
 ];
-
-const DEFAULT_SUPERVISION_TYPES = [
-  "all",
-  "dual",
-  "parole",
-  "probation",
-  "external_unknown",
-];
-
-const DEFAULT_SUPERVISION_LEVELS = ["all"];
 
 const DEFAULT_VIOLATION_TYPES = [
   "absconded",
   "all",
   "elec_monitoring",
   "escaped",
-  "felony",
   "high_tech",
+  "law",
   "low_tech",
   "med_tech",
-  "misdemeanor",
   "municipal",
   "no_violation_type",
   "substance_abuse",
   "technical",
 ];
 
-const DEFAULT_RISK_LEVEL = [
-  "high",
-  "low",
-  "medium",
-  "not_assessed",
-  "very_high",
-];
+const DEFAULT_SUPERVISION_TYPES = ["all"];
 
-const US_MO_DIMENSIONS = {
+const DEFAULT_CHARGE_CATEGORIES = ["all"];
+
+const DEFAULT_RISK_LEVEL = ["high", "low", "medium", "not_assessed"];
+
+const US_PA_DIMENSION_VALUES = {
   charge_category: DEFAULT_CHARGE_CATEGORIES,
   gender: DEFAULT_GENDERS,
   metric_period_months: DEFAULT_METRIC_PERIOD_MONTHS,
@@ -78,4 +67,4 @@ const US_MO_DIMENSIONS = {
   violation_type: DEFAULT_VIOLATION_TYPES,
 };
 
-module.exports = { US_MO_DIMENSIONS };
+module.exports = { US_PA_DIMENSION_VALUES };
